split shared, remove Ptr() methods
