+++ shared.go
@@ -1,65 +1,14 @@
 package starr
 
 import (
-	"crypto/tls"
 	"encoding/json"
-	"net/http"
 	"net/url"
 	"strconv"
 	"strings"
 	"time"
-
-	"golift.io/starr/debuglog"
 )
 
 /* This file contains shared structs or constants for all the *arr apps. */
-
-// App can be used to satisfy a context value key.
-// It is not used in this library; provided for convenience.
-type App string
-
-// These constants are just here for convenience.
-const (
-	Emby     App = "Emby"
-	Lidarr   App = "Lidarr"
-	Plex     App = "Plex"
-	Prowlarr App = "Prowlarr"
-	Radarr   App = "Radarr"
-	Readarr  App = "Readarr"
-	Sonarr   App = "Sonarr"
-	Whisparr App = "Whisparr"
-)
-
-// String turns an App name into a string.
-func (a App) String() string {
-	return string(a)
-}
-
-// Lower turns an App name into a lowercase string.
-func (a App) Lower() string {
-	return strings.ToLower(string(a))
-}
-
-// Client returns the default client, and is used if one is not passed in.
-func Client(timeout time.Duration, verifySSL bool) *http.Client {
-	return &http.Client{
-		Timeout: timeout,
-		CheckRedirect: func(_ *http.Request, _ []*http.Request) error {
-			return http.ErrUseLastResponse
-		},
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: !verifySSL}, //nolint:gosec
-		},
-	}
-}
-
-// ClientWithDebug returns an http client with a debug logger enabled.
-func ClientWithDebug(timeout time.Duration, verifySSL bool, logConfig debuglog.Config) *http.Client {
-	client := Client(timeout, verifySSL)
-	client.Transport = debuglog.NewLoggingRoundTripper(logConfig, client.Transport)
-
-	return client
-}
 
 // CalendarTimeFilterFormat is the Go time format the calendar expects the filter to be in.
 const CalendarTimeFilterFormat = "2006-01-02T03:04:05.000Z"
@@ -235,6 +184,13 @@ func (o *QueueDeleteOpts) Values() url.Values {
 	return params
 }
 
+// FormatItem is part of a quality profile.
+type FormatItem struct {
+	Format int64  `json:"format"`
+	Name   string `json:"name"`
+	Score  int64  `json:"score"`
+}
+
 // PlayTime is used in at least Sonarr, maybe other places.
 // Holds a string duration converted from hh:mm:ss.
 type PlayTime struct {
@@ -242,12 +198,7 @@ type PlayTime struct {
 	time.Duration
 }
 
-// FormatItem is part of a quality profile.
-type FormatItem struct {
-	Format int64  `json:"format"`
-	Name   string `json:"name"`
-	Score  int64  `json:"score"`
-}
+var _ json.Unmarshaler = (*PlayTime)(nil)
 
 // UnmarshalJSON parses a run time duration in format hh:mm:ss.
 func (d *PlayTime) UnmarshalJSON(b []byte) error {
@@ -275,8 +226,6 @@ func (d *PlayTime) MarshalJSON() ([]byte, error) {
 	return []byte(`"` + d.Original + `"`), nil
 }
 
-var _ json.Unmarshaler = (*PlayTime)(nil)
-
 // ApplyTags is an enum used as an input for Bulk editors, and perhaps other places.
 type ApplyTags string
 
@@ -287,11 +236,6 @@ const (
 	TagsRemove  ApplyTags = "remove"
 	TagsReplace ApplyTags = "replace"
 )
-
-// Ptr returns a pointer to an apply tags value. Useful for a BulkEdit struct.
-func (a ApplyTags) Ptr() *ApplyTags {
-	return &a
-}
 
 // TimeSpan is part of AudioTags and possibly used other places.
 type TimeSpan struct {
@@ -306,33 +250,6 @@ type TimeSpan struct {
 	TotalMilliseconds int64 `json:"totalMilliseconds"`
 	TotalMinutes      int64 `json:"totalMinutes"`
 	TotalSeconds      int64 `json:"totalSeconds"`
-}
-
-// Itoa converts an int64 to a string.
-// Deprecated: Use starr.Str() instead.
-func Itoa(v int64) string {
-	return Str(v)
-}
-
-// Str converts numbers and booleans to a string.
-func Str[I int | int64 | float64 | bool](val I) string {
-	const (
-		base10 = 10
-		bits64 = 64
-	)
-
-	switch val := interface{}(val).(type) {
-	case int:
-		return strconv.Itoa(val)
-	case bool:
-		return strconv.FormatBool(val)
-	case int64:
-		return strconv.FormatInt(val, base10)
-	case float64:
-		return strconv.FormatFloat(val, 'f', -1, bits64)
-	default:
-		return ""
-	}
 }
 
 // Protocol used to download media. Comes with enum constants.
